+++ owl-back/src/main/java/com/project/owlback/codereview/repository/CodeHistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CodeHistoryRepository extends JpaRepository<CodeHistory, Integer> {
+public interface CodeHistoryRepository extends JpaRepository<CodeHistory, Long> {
 }
